# docs/rebuild/openapi-baseline.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Tradier Bot API",
  "description": "Standalone backend for three worlds on one process: the Tradier options desk (balance, chain preview, managed positions with TP/SL, the opening-range and A/B auto-traders, the HOT scan and the options-flow board), the 36 Trade Desk that fronts it on a phone, and the Bot Station that launches, supervises and settles the vendored Kalshi bots (btc15, btc60, sports, parlay, gold15, silver15, oil15). Plus the super_research signal engines and the SPY/QQQ/SPX level watcher they read. SQLite persistence, per-user credential folders, paper-only by default.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/super/gex0dte/refresh": {
   "post": {
    "tags": [
     "super-research"
    ],
    "summary": "Refresh Gex0Dte",
    "operationId": "refreshGex0dte",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Gex0dteRefresh"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Refreshgex0Dte"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/super/gex0dte/heartbeat": {
   "post": {
    "tags": [
     "super-research"
    ],
    "summary": "Gex0Dte Heartbeat",
    "description": "Record that a push cycle happened, whatever its outcome.\n\nStrictly separate from the data path: it never fetches the vendor, never\nstores a snapshot and never fills an hour slot. Routing liveness through\n/refresh would make the SERVER call getgamma on every failed cycle \u2014 the\none thing this design does not do \u2014 and would stamp a stalled feed as\nfresh.",
    "operationId": "gex0dtePusherHeartbeat",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/HeartbeatIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Gex0Dtepusherheartbeat"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Gex0dteRefresh": {
    "properties": {
     "payload": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Payload"
     },
     "ticker": {
      "type": "string",
      "title": "Ticker",
      "default": "SPY"
     },
     "client": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Client"
     }
    },
    "type": "object",
    "title": "Gex0dteRefresh",
    "description": "Give it a chain captured from a getgamma.io tab, or nothing to make the\nserver try the vendor directly. No credentials either way \u2014 the endpoint\nneeds none; its edge simply refuses non-browser clients."
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "HeartbeatIn": {
    "properties": {
     "session": {
      "type": "string",
      "title": "Session",
      "default": "?"
     },
     "seq": {
      "type": "integer",
      "title": "Seq",
      "default": 0
     },
     "ok": {
      "type": "boolean",
      "title": "Ok",
      "default": false
     },
     "reason": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason"
     },
     "wall": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Wall"
     },
     "mono": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Mono"
     }
    },
    "type": "object",
    "title": "HeartbeatIn",
    "description": "One push cycle reported by the browser pusher, successful or not."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}